(structure refactor): header
