+++ src/components/Header/AppHeaderUser.tsx
@@ -1,0 +1,149 @@
+import { useWeb3React } from "@web3-react/core";
+import AddressDropdown from "../AddressDropdown/AddressDropdown";
+import { ConnectWalletButton } from "../Common/Button";
+import React, { useCallback, useEffect } from "react";
+import { HeaderLink } from "./HeaderLink";
+import connectWalletImg from "../../img/ic_wallet_24.svg";
+
+import "./Header.css";
+import { isHomeSite, ARBITRUM, AVALANCHE, getAccountUrl, getChainName,  switchNetwork, useChainId } from "../../lib/legacy";
+import {ARBITRUM_TESTNET, isDevelopment} from "../../lib/legacy";
+import cx from "classnames";
+import {Trans} from "@lingui/macro";
+import NetworkDropdown from "../NetworkDropdown/NetworkDropdown";
+import LanguagePopupHome from "../NetworkDropdown/LanguagePopupHome";
+
+type Props = {
+  openSettings: () => void;
+  small?: boolean;
+  setWalletModalVisible: (visible: boolean) => void;
+  disconnectAccountAndCloseSettings: () => void;
+  redirectPopupTimestamp: number;
+  showRedirectModal: (to: string) => void;
+};
+
+export function AppHeaderUser({
+ openSettings,
+ small,
+ setWalletModalVisible,
+ disconnectAccountAndCloseSettings,
+ redirectPopupTimestamp,
+ showRedirectModal
+}: Props) {
+  const { chainId } = useChainId();
+  const { active, account } = useWeb3React();
+  const showConnectionOptions = !isHomeSite();
+
+  const networkOptions = [
+    {
+      label: getChainName(ARBITRUM),
+      value: ARBITRUM,
+      icon: "ic_arbitrum_24.svg",
+      color: "#264f79",
+    },
+    {
+      label: getChainName(AVALANCHE),
+      value: AVALANCHE,
+      icon: "ic_avalanche_24.svg",
+      color: "#E841424D",
+    },
+  ];
+  if (isDevelopment()) {
+    networkOptions.push({
+      label: getChainName(ARBITRUM_TESTNET),
+      value: ARBITRUM_TESTNET,
+      icon: "ic_arbitrum_24.svg",
+      color: "#264f79",
+    });
+  }
+
+  useEffect(() => {
+    if (active) {
+      setWalletModalVisible(false);
+    }
+  }, [active, setWalletModalVisible]);
+
+  const onNetworkSelect = useCallback(
+    (option) => {
+      if (option.value === chainId) {
+        return;
+      }
+      return switchNetwork(option.value, active);
+    },
+    [chainId, active]
+  );
+
+  const selectorLabel = getChainName(chainId);
+
+  if (!active) {
+    return (
+      <div className="App-header-user">
+        <div className={cx("App-header-trade-link", { "homepage-header": isHomeSite() })}>
+          <HeaderLink
+              className="default-btn"
+              to="/trade"
+              redirectPopupTimestamp={redirectPopupTimestamp}
+              showRedirectModal={showRedirectModal}
+          >
+            <Trans>Trade</Trans>
+          </HeaderLink>
+        </div>
+
+        {showConnectionOptions ? (
+          <>
+            <ConnectWalletButton onClick={() => setWalletModalVisible(true)} imgSrc={connectWalletImg}>
+              {small ? <Trans>Connect</Trans> : <Trans>Connect Wallet</Trans>}
+            </ConnectWalletButton>
+            <NetworkDropdown
+              small={small}
+              networkOptions={networkOptions}
+              selectorLabel={selectorLabel}
+              onNetworkSelect={onNetworkSelect}
+              openSettings={openSettings}
+            />
+          </>
+        ) : (
+          <LanguagePopupHome />
+        )}
+      </div>
+    );
+  }
+
+  const accountUrl = getAccountUrl(chainId, account);
+
+  return (
+    <div className="App-header-user">
+      <div className="App-header-trade-link">
+        <HeaderLink
+          className="default-btn"
+          to="/trade"
+          redirectPopupTimestamp={redirectPopupTimestamp}
+          showRedirectModal={showRedirectModal}
+        >
+          <Trans>Trade</Trans>
+        </HeaderLink>
+      </div>
+
+      {showConnectionOptions ? (
+        <>
+          <div className="App-header-user-address">
+            <AddressDropdown
+              account={account}
+              accountUrl={accountUrl}
+              disconnectAccountAndCloseSettings={disconnectAccountAndCloseSettings}
+            />
+          </div>
+          <NetworkDropdown
+            small={small}
+            networkOptions={networkOptions}
+            selectorLabel={selectorLabel}
+            onNetworkSelect={onNetworkSelect}
+            openSettings={openSettings}
+          />
+        </>
+      ) : (
+        <LanguagePopupHome />
+      )}
+    </div>
+  );
+}
